Support displaying multiple variants for a template part
Allows multiple data entries to be registered for a template part
through the header comments. Any entries included in @data or
@dataProvider tags will be merged together and used to display different
variants.

(If a header is present but no data tag is provided, the template part
will not render. If no header is present, the template part will render
once with an empty data array (ie, just making use of global values from
the latest post on the site.)